# swagger.yaml
openapi: 3.0.3
info:
  title: Site e-commerce
  description: >
    Some useful links:

    - [Lien Github du code
    source](https://github.com/DavyHaYehoudi/e-commerce-Cindy)

    - [The source API definition for the Pet
    Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  contact:
    email: davy.programa@gmail.com
  version: 1.0.0
servers:
  - url: https://ateliernoralya.com
...
paths:
...
  /products/{productsId}:
    get:
      tags:
        - product
      summary: Get a product by ID
      description: This route retrieves details of a specific product based on its unique identifier (productsId). Information about the product, such as its name, description, price, etc., is returned in the response.
      operationId: getProductById
      parameters:
        - name: productsId
          in: path
          description: ID of products that needs to be fetched
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductCreateResponse'
        '404':
          description: The product does not exist.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Le produit n'existe pas.
    patch:
      tags:
        - product
      summary: Modifying a product
      description: This route is used by administrators to modify the details of a specific product identified by its unique identifier (productsId). The fields to be updated, such as name, description, price, etc., are provided in the body of the PATCH request.
      operationId: editProduct
      parameters:
        - name: productsId
          in: path
          description: ID of product that needs to be fetched
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Successful updated product
          content:
           application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateResponse'
        '403':
          description: Unauthorized - Admin is not authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Accès refusé. Vous n'êtes pas un administrateur.
        '404':
          description: The product does not exist.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Le produit n'existe pas.
      security:
        - api_key: []
    delete:
      tags:
        - product
      summary: Delete product
      description: This route is used by administrators to remove a specific product from the catalogue based on its unique identifier (productsId).
      operationId: deleteProduct
      parameters:
        - name: productsId
          in: path
          description: ID of the order that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful deleted product
          content:
            application/json:
              schema:
                type: object
                properties:
                  productId:
                    type: string
        '403':
          description: Unauthorized - Admin is not authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Accès refusé. Vous n'êtes pas un administrateur.
        '404':
          description: The ressource does not exist.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Le produit n'existe pas.
      security:
        - api_key: []
...
  /promocodes/{promocodeId}:
    delete:
      tags:
        - promocode
      summary: Delete a promocode 
      description: This route, reserved for administrators, allows deleting a specific promotional code. The promotional code to be deleted is identified by its unique code.
      operationId: deletePromocode
      parameters:
        - name: promocodeId
          in: path
          description: ID of the promocode that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful deleted promocode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromocodeDeleteResponse'
        '403':
          description: Unauthorized - Admin is not authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Accès refusé. Vous n'êtes pas un administrateur.
        '404':
          description: The ressource does not exist.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Le code promo n'existe pas.
      security:
        - api_key: []
...
  /tags/{tagId}:
    put:
      tags:
        - tag
      summary: Update a tag
      description: This route, reserved for administrators, allows updating an existing tag. Details of the tag to be updated, such as its name and the products it is associated with, are provided in the PUT request body.
      operationId: updateTag
      parameters:
        - name: tagId
          in: path
          description: ID of the tag that needs to be updated
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagCreate'
      responses:
        '200':
          description: Successful updated tag
          content:
           application/json:
            schema:
              $ref: '#/components/schemas/Tag'
        '403':
          description: Unauthorized - Admin is not authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Accès refusé. Vous n'êtes pas un administrateur.
        '404':
          description: The ressource does not exist.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Le tag n'existe pas.
      security:
        - api_key: []
    delete:
      tags:
        - tag
      summary: Delete a tag
      description: This route, reserved for administrators, allows deleting a specific tag.
      operationId: deleteTag
      parameters:
        - name: tagId
          in: path
          description: ID of the tag that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful deleted tag
          content:
           application/json:
            schema:
              $ref: '#/components/schemas/Tag'
        '403':
          description: Unauthorized - Admin is not authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Accès refusé. Vous n'êtes pas un administrateur.
        '404':
          description: The ressource does not exist.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Le tag n'existe pas.
      security:
        - api_key: []
components:
  schemas:
...
    Product:
      type: object
      properties:
        name:
          type: string
          example: Boucles d'oreilles
        collection:
          type: string
          example: 65fafdf6ab75220625f55db1
        category:
          type: string
          example: 65fafe1eab75220625f55db5
        tags:
          type: array
          items:
            type: string
            example: 65fafea0ab75220625f55dc4
        secondary_images:
          type: array
          items: 
            type: string
            example: bo.png
        main_description:
          type: string
          example: >-
            Magnifiques boucles d'oreilles pour des occasions uniques et qui
            vous mettront en valeur.
        createdAt:
          type: string
          example: "2024-03-21T17:50:38.100Z"
        updatedAt:
          type: string
          example: "2024-03-21T17:50:38.100Z"
        materials:
          type: array
          items:
            type: object
            properties:
              _id:
                type: string
                example: 65f2d2df4db605d020200620
              pricing:
                type: object
                properties:
                  currentPrice:
                    type: string
                    example: 15
                  oldPrice:
                    type: string
                    example: 30
              promotion:
                type: object
                properties:
                  amount:
                    type: integer
                    format: int16
                    example: 50
                  startDate:
                    type: string
                    example: 2024-04-01T00:00:00.000
                  endDate:
                    type: string
                    example: 2024-05-01T00:00:00.000Z
              main_image:
                type: string
                example: col.png
              untilNew:
                type: string
                example: 2024-04-15T00:00:00.000Z
              stock:
                type: integer
                format: int16
                example: 7
              isActive:
                type: boolean
                example: true
...
    ProductCreateResponse:
      type: object
      properties:
        _id: 
          type: string
        name:
          type: string
          example: Boucles d'oreilles
        collection:
          type: string
          example: 65fafdf6ab75220625f55db1
        category:
          type: string
          example: 65fafe1eab75220625f55db5
        tags:
          type: array
          items:
            type: string
            example: 65fafea0ab75220625f55dc4
        secondary_images:
          type: array
          items: 
            type: string
            example: bo.png
        main_description:
          type: string
          example: >-
            Magnifiques boucles d'oreilles pour des occasions uniques et qui
            vous mettront en valeur.
        materials:
          type: array
          items:
            type: object
            required:
              - _id
              - pricing
              - main_image
              - untilNew
              - stock
            properties:
              _id:
                type: string
                example: 65f2d2df4db605d020200620
              pricing:
                type: object
                properties:
                  currentPrice:
                    type: string
                    example: 15
                  oldPrice:
                    type: string
                    example: 30
              promotion:
                type: object
                properties:
                  amount:
                    type: integer
                    format: int16
                    example: 50
                  startDate:
                    type: string
                    example: 2024-04-01T00:00:00.000
                  endDate:
                    type: string
                    example: 2024-05-01T00:00:00.000Z
              main_image:
                type: string
                example: col.png
              untilNew:
                type: string
                example: 2024-04-15T00:00:00.000Z
              stock:
                type: integer
                format: int16
                example: 7
...
    Tag:
      type: array
      items:
        type: object
        properties:
          _id:
            type: string
            example: 2884d22b45a0fb3fd4a75p2
          name:
            type: string
            example: Elégance
    TagCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: Fantaisie
...
    PromocodeDeleteResponse:
      type: object
      properties:
        promocodeId:
          type: string
...
  securitySchemes:
    jwt_auth:
      type: apiKey
      in: header
      name: Authorization
      description: JWT authorization header using the Bearer scheme.
    api_key:
      type: apiKey
      name: api_key
      in: header